Add pgBackRest v.2.45.
Update pgBackRest to v2.45 for docker images and builds.
Decrease value from 3 to 1 for pgbackrest_backup_repo_size_bytes metric in e2e-tests. This is necessary because the pgbackrest_backup_repo_size_bytes metric is not available for block incremental backups (see backups to repo 2 in prepare_e2e.sh).