Set default value for metrics.
For metrics:
* pgbackrest_backup_repo_size_bytes
* pgbackrest_backup_repo_size_map_bytes
* pgbackrest_backup_repo_delta_map_bytes
* pgbackrest_backup_last_repo_size_bytes
* pgbackrest_backup_last_repo_size_map_bytes
* pgbackrest_backup_last_repo_delta_map_bytes

If value is absent, metric will be set to 0.
It's necessary to avoid flapping time series.

It is much more reasonable to pass the value 0 than to handle the event of the absence metrics. It's necessary to avoid flapping time series. This is especially important for *_last_* metrics.