Fix incorrect latest metrics when filtering backup types.
When `--backrest.backup-type` flag is specified, if parameter value is `incr` or `diff`, all `"pgbackrest_backup_last_*` metrics will not be collected.
When parameter value is `full`, the metrics will be as if the last backup was `full` (i.e. the same for `diff` and `incr`).<br>

Also removed outdate info about  `pgbackrest_backup_annptations` metric.